# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TinyClaw API",
    "version": "1",
    "description": "HTTP API for the TinyClaw personal AI assistant."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4310",
      "description": "Local dev server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "ListProfilesResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "ProfileResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "CreateProfileRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "SoulStatusResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "SoulStackResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "InitSoulResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "UpdateSoulFileRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "ListKnowledgeBaseResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "UploadKnowledgeBaseResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "UploadKnowledgeBaseRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "DeleteKnowledgeBaseResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "ImageAttachment": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
...
    },
...
  },
  "paths": {
...
    "/v1/profiles": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "List bot profiles",
        "operationId": "listProfiles",
        "responses": {
          "200": {
            "description": "Profile list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListProfilesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Create a bot profile",
        "operationId": "createProfile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "Get a bot profile",
        "operationId": "getProfile",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profiles"
        ],
        "summary": "Update a bot profile",
        "operationId": "updateProfile",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "summary": "Delete a bot profile",
        "operationId": "deleteProfile",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Profile deleted"
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/soul": {
      "get": {
        "tags": [
          "Soul",
          "Profiles"
        ],
        "summary": "Get soul status for a profile",
        "operationId": "getProfileSoulStatus",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "contents",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Soul status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoulStatusResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/soul/stack": {
      "get": {
        "tags": [
          "Soul",
          "Profiles"
        ],
        "summary": "Get soul stack contents for a profile",
        "operationId": "getProfileSoulStack",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Soul stack",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoulStackResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/soul/init": {
      "post": {
        "tags": [
          "Soul",
          "Profiles"
        ],
        "summary": "Initialize soul templates for a profile",
        "operationId": "initProfileSoul",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Soul initialized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitSoulResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/soul/files/{fileKey}": {
      "put": {
        "tags": [
          "Soul",
          "Profiles"
        ],
        "summary": "Write a profile soul file",
        "operationId": "writeProfileSoulFile",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "soul",
                "style",
                "skill",
                "memory"
              ]
            },
            "required": true,
            "name": "fileKey",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSoulFileRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "File saved"
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/knowledge-base": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "List knowledge base documents for a profile",
        "operationId": "listKnowledgeBase",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Knowledge base documents",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListKnowledgeBaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Profiles"
        ],
        "summary": "Upload a knowledge base document",
        "operationId": "uploadKnowledgeBaseDocument",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadKnowledgeBaseRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Uploaded knowledge base document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadKnowledgeBaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/knowledge-base/{documentId}": {
      "delete": {
        "tags": [
          "Profiles"
        ],
        "summary": "Delete a knowledge base document",
        "operationId": "deleteKnowledgeBaseDocument",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "documentId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted knowledge base document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteKnowledgeBaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}/avatar": {
      "get": {
        "tags": [
          "Profiles"
        ],
        "summary": "Get a profile avatar image",
        "operationId": "getProfileAvatar",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Profile avatar image",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Profiles"
        ],
        "summary": "Upload a profile avatar",
        "operationId": "uploadProfileAvatar",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageAttachment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile with updated avatar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Profiles"
        ],
        "summary": "Delete a profile avatar",
        "operationId": "deleteProfileAvatar",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "profileId",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Avatar deleted"
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}